+++ pyproject.toml
@@ -7,7 +7,7 @@ name = "cluster-snek"
 version = "0.1.1"
 description = "Professional-grade Kubernetes GitOps automation with comprehensive deployment modes"
 requires-python = ">=3.12"
-dependencies = ["PyYAML"]
+dependencies = ["PyYAML", "GitPython"]
 
 [project.optional-dependencies]
 test = ["pytest"]
